CLN: Create tests/window/moments/conftest for specific fixtures
